requirements.txt
PySide6-QtAds
PySide6-essentials
vtk
matplotlib~=3.7.1
numpy<2.0
scipy~=1.10.1
IPython~=8.12.2
rapidfuzz
pytest~=7.3.2
pytest-cov
pytest-runner
xarray~=2023.12.0
PyYAML~=6.0
Pillow~=10.3.0